# src/main/java/io/bosch/enterprisesystems/api/dto/EnterpriseSystemsCreateRequest.java
package io.bosch.enterprisesystems.api.dto;

import lombok.Builder;

/**
 * EnterpriseSystemsRequest
 */
@Builder
public class EnterpriseSystemsCreateRequest {

    private String name;
    private String endpointUrl;
    private String authMethod;

}